fix : fixed postcode component to typescript
